親SOMの潜在空間に観測データを表示して可視化するtutorialを追加した
