+++ tutorials/som2/curve1.py
@@ -1,0 +1,95 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.animation as anim
+
+from libs.models.som2 import SOM2
+
+if __name__ == "__main__":
+    seed = 2
+    np.random.seed(seed)
+    nb_epoch = 30
+    n_class = 6
+    n_sample = 400
+    Dim = 2
+    parent_latent_dim = 2
+    child_latent_dim = 1
+    parent_resolution = 8
+    child_resolution = 100
+    pCluster_num = parent_resolution ** parent_latent_dim
+    cCluster_num = child_resolution ** child_latent_dim
+    parent_sigma_max = 2.0
+    parent_sigma_min = 0.3
+    child_sigma_max = 2.0
+    child_sigma_min = 0.15
+    parent_tau = nb_epoch
+    child_tau = nb_epoch
+    interval = 500
+
+    # ZetaとZの初期値を生成
+    if parent_latent_dim == 2:
+        pZ = np.random.normal(size=(n_class, parent_latent_dim), loc=0, scale=0.01)
+    else:
+        pZ = np.random.normal(size=(n_class, 1), loc=0.0, scale=0.01)
+
+    if child_latent_dim == 2:
+        cZ = np.random.normal(size=(n_sample, child_latent_dim), loc=0, scale=0.01)
+    else:
+        cZ = np.random.normal(size=(n_sample, 1), loc=0.0, scale=0.01)
+
+    # データ生成
+    Datasets = np.zeros((n_class, n_sample, Dim))
+    x = np.random.rand(n_class, n_sample) * 2 - 1
+    x = np.sort(x)
+    y = [0.5 * x[0], -0.5 * x[1], 0.5 * x[2] ** 2, -0.5 * (x[3] ** 2), x[4] ** 3 - 0.5 * x[4], -x[5] ** 3 + 0.5 * x[5]]
+    for n, (a, b) in enumerate(zip(x, y)):
+        Datasets[n, :, 0] = a
+        Datasets[n, :, 1] = b
+
+    model = SOM2(Datasets, parent_latent_dim, child_latent_dim, parent_resolution, child_resolution,
+                 parent_sigma_max, child_sigma_max, parent_sigma_min, child_sigma_min,
+                 parent_tau, child_tau, pZ, cZ)
+    model.fit(nb_epoch)
+
+    cY = model.history["cY"]
+    pY = model.history["pY"]
+    cZ = model.history["cZ"]
+    pZ = model.history["pZ"]
+    pZeta = model.history["pZeta"]
+    bmu = model.history["bmu"]
+
+    fig, axes = plt.subplots(parent_resolution, parent_resolution, figsize=(7, 7))
+
+    def update(epoch):
+        for i in range(parent_resolution):
+            for j in range(parent_resolution):
+                axes[i, j].cla()
+
+        unique_bmu = np.unique(bmu[epoch])
+        list_bmu = [0]*pCluster_num
+        for i in unique_bmu:
+            list_bmu[int(i)] = 1
+
+        t = 0
+        for i in range(parent_resolution):
+            for j in range(parent_resolution):
+                if list_bmu[t] == 1:
+                    axes[i, j].scatter(pY[epoch, t, :, 0], pY[epoch, t, :, 1], s=5, color='r', label='bmu')
+                    axes[i, j].set_facecolor('k')
+                    axes[i, j].legend(loc=2, prop={'size': 5})
+                else:
+                    axes[i, j].scatter(pY[epoch, t, :, 0], pY[epoch, t, :, 1], s=5, color='grey')
+                    axes[i, j].set_facecolor('w')
+                t += 1
+                axes[i, j].set_xticks([])
+                axes[i, j].set_yticks([])
+                axes[i, j].set_xlim(-1, 1)
+                axes[i, j].set_ylim(-1, 1)
+
+        fig.suptitle("epoch {}/{} latent space(parent)".format((epoch + 1), nb_epoch), fontsize=10)
+
+    ani = anim.FuncAnimation(fig, update, interval=interval, frames=nb_epoch, repeat=False)
+    ani.save("SOM2.gif", writer='pillow')
+    # plt.show()
